requirements.txt
streamlit
pyyaml
aiohttp
nest_asyncio
asyncio
agno
openai
anthropic
google-genai
pre-commit
groq
ibm-watsonx-ai
agentql
psycopg2-binary
pypdf
langchain
langchain-community
pgvector
psycopg
